requirements.txt
# FaceCV Requirements - Production
# Python 3.10+

# Core dependencies
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
scipy==1.11.4

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database drivers
SQLAlchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
python-dotenv==1.0.0

# Face recognition
insightface==0.7.3
deepface==0.0.79
onnxruntime==1.16.3

# Vector database
chromadb==0.4.18

# Utilities
scikit-learn==1.3.2
requests==2.31.0
pyyaml==6.0.1

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Performance
psutil==5.9.6
py-cpuinfo==9.0.0